relocated donate javascript and imported it in app.js
